renamed vendor -> store, store -> shop
