+++ hamza-client/src/modules/products/components/product-group-store/index.tsx
@@ -8,49 +8,43 @@ import {
     Grid,
     GridItem,
     Flex,
-    Button,
 } from '@chakra-ui/react';
 import axios from 'axios';
 import { useQuery } from '@tanstack/react-query';
-import { formatCryptoPrice } from '@lib/util/get-product-price';
 import { useCustomerAuthStore } from '@store/customer-auth/customer-auth';
-import ProductCardStore from '@modules/shop/components/product-card';
-import useStorePage from '@store/store-page/store-page';
+import ProductCardHome from '../product-group-home/component/home-product-card';
+import { formatCryptoPrice } from '@lib/util/get-product-price';
+import useVendor from '@store/store-page/vendor';
 
-const ProductCardGroup = () => {
+type Props = {
+    storeName: string;
+    filterByRating?: string | null;
+};
+
+const ProductCardGroup = ({ storeName }: Props) => {
+    // get preferred currency
     const { preferred_currency_code } = useCustomerAuthStore();
-    const { categorySelect } = useStorePage();
-    const [visibleProductsCount, setVisibleProductsCount] = useState(15); // State to manage visible products count (4 rows, 16 items)
 
-    console.log('This is the categorySelect from Store Page', categorySelect);
-    //TODO: MOVE TO INDEX.TS
-    // Get products from vendor
-    const multiUrl = `${process.env.NEXT_PUBLIC_MEDUSA_BACKEND_URL || 'http://localhost:9000'}/custom/product/filter?category_name=${categorySelect}&price_hi=${5000000}&price_lo=${0}`;
+    const { categorySelect } = useVendor();
 
-    //TODO: MOVE TO INDEX.TS
-    // Get products from vendor
+    const url = `${process.env.NEXT_PUBLIC_MEDUSA_BACKEND_URL || 'http://localhost:9000'}/custom/store/products/category-name?store_name=${storeName}&category_name=${categorySelect}`;
+
     const { data, error, isLoading } = useQuery(
-        ['categories', categorySelect], // Use a unique key here to identify the query
+        ['products', categorySelect],
         async () => {
-            const response = await axios.get(multiUrl);
-            return response.data; // Return the data from the response
+            console.log('Fetching data from URL:', url);
+            const response = await axios.get(url);
+            console.log('Data fetched:', response.data);
+            return response.data;
         }
     );
 
-    const productsAll = data?.products || [];
-
-    const handleViewMore = () => {
-        // Increase the visible products count by 16 (4 rows of 4 products)
-        setVisibleProductsCount((prevCount) => prevCount + 15);
-    };
-
-    const visibleProducts = productsAll.slice(0, visibleProductsCount);
+    // Handle products based on category
+    const productsAll = data;
 
     if (isLoading) {
         return (
             <Flex
-                mt={{ base: '0', md: '1rem' }}
-                mb={'4rem'}
                 maxW={'1280px'}
                 width="100%"
                 mx="auto"
@@ -63,11 +57,11 @@ const ProductCardGroup = () => {
                     width="100%"
                     templateColumns={{
                         base: 'repeat(2, 1fr)',
-                        lg: 'repeat(3, 1fr)',
+                        lg: 'repeat(4, 1fr)',
                     }}
                     gap={{ base: '4', md: '25.5px' }}
                 >
-                    {Array.from({ length: 9 }).map((_, index) => (
+                    {Array.from({ length: 8 }).map((_, index) => (
                         <GridItem
                             key={index}
                             minHeight="243.73px"
@@ -96,85 +90,73 @@ const ProductCardGroup = () => {
     }
 
     return (
-        <Box maxW={'941px'} w="100%">
+        <Flex
+            mb={'4rem'}
+            maxW={'1280px'}
+            width="100%"
+            mx="auto"
+            justifyContent={'center'}
+            alignItems={'center'}
+        >
             <Grid
-                mt={{ base: '0px', md: '2rem' }}
-                mx={{ base: '1rem', md: '0' }}
+                maxWidth={'1256.52px'}
+                mx="1rem"
+                width="100%"
                 templateColumns={{
                     base: 'repeat(2, 1fr)',
-                    md: 'repeat(3, 1fr)',
+                    lg: 'repeat(4, 1fr)',
                 }}
-                gap={{ base: 4, md: 7 }}
+                gap={{ base: '4', md: '25.5px' }}
             >
-                {productsAll?.length &&
-                    visibleProducts.map((product: any, index: number) => {
-                        // Extract product details
-                        const variant = product?.variants[0]; // Assuming you want the first variant
-                        const productPricing =
-                            variant?.prices?.find(
-                                (price: any) =>
-                                    price.currency_code ===
-                                    (preferred_currency_code ?? 'usdc')
-                            )?.amount ||
-                            variant?.prices?.[0]?.amount ||
-                            0; // Get the price for the preferred currency or fallback to the first price
+                {productsAll.map((product: any, index: number) => {
+                    const variant = product.variants[0];
+                    const productPricing =
+                        variant?.prices?.find(
+                            (price: any) =>
+                                price.currency_code === (preferred_currency_code ?? 'usdc')
+                        )?.amount ||
+                        variant?.prices?.[0]?.amount ||
+                        0;
+                    const formattedPrice = formatCryptoPrice(
+                        productPricing ?? 0,
+                        (preferred_currency_code ?? 'usdc') as string
+                    );
 
-                        const formattedPrice = formatCryptoPrice(
-                            productPricing ?? 0,
-                            (preferred_currency_code ?? 'usdc') as string
-                        );
-
-                        return (
-                            <GridItem
+                    return (
+                        <GridItem
+                            key={index}
+                            //   maxW={'295px'}
+                            minHeight={'243.73px'}
+                            height={{ base: '100%', md: '399px' }}
+                            width="100%"
+                        >
+                            <ProductCardHome
                                 key={index}
-                                minHeight={'243.73px'}
-                                height={{ base: '100%', md: '399px' }}
-                                width="100%"
-                            >
-                                <ProductCardStore
-                                    key={index}
-                                    productHandle={product?.handle ?? ''}
-                                    reviewCount={product?.review}
-                                    totalRating={10}
-                                    variantID={variant?.id ?? ''}
-                                    countryCode={product?.origin_country ?? ''}
-                                    productName={product?.title ?? ''}
-                                    productPrice={formattedPrice ?? ''}
-                                    currencyCode={
-                                        preferred_currency_code || 'usdc'
-                                    }
-                                    imageSrc={product?.thumbnail ?? ''}
-                                    hasDiscount={product?.is_giftcard ?? false}
-                                    discountValue={product?.discountValue ?? ''}
-                                    productId={product?.id ?? ''}
-                                    inventory={variant?.inventory_quantity ?? 0}
-                                    allow_backorder={variant?.allow_backorder}
-                                    storeId={product?.store_id ?? ''}
-                                />
-                            </GridItem>
-                        );
-                    })}
+                                productHandle={product.handle}
+                                reviewCount={0}
+                                totalRating={0}
+                                variantID={'0'}
+                                countryCode={product.countryCode}
+                                productName={product.title}
+                                productPrice={formattedPrice}
+                                currencyCode={preferred_currency_code ?? 'usdc'}
+                                imageSrc={product.thumbnail}
+                                hasDiscount={product.is_giftcard}
+                                discountValue={product.discountValue}
+                                productId={product.id}
+                                inventory={
+                                    product.variants[0]?.inventory_quantity
+                                }
+                                allow_backorder={
+                                    product.variants[0]?.allow_backorder
+                                }
+                                storeId={product.store_id}
+                            />
+                        </GridItem>
+                    );
+                })}
             </Grid>
-
-            {/* Show the "View More" button only if there are more products to display */}
-            <Flex>
-                {visibleProductsCount < productsAll.length && (
-                    <Button
-                        mt="2rem"
-                        mx="auto"
-                        alignSelf={'center'}
-                        textAlign={'center'}
-                        onClick={handleViewMore}
-                        variant="solid"
-                        borderRadius={'full'}
-                        backgroundColor={'white'}
-                        color="black"
-                    >
-                        Show More
-                    </Button>
-                )}
-            </Flex>
-        </Box>
+        </Flex>
     );
 };
