+++ hamza-client/src/modules/shop/index.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { Flex } from '@chakra-ui/react';
 import SideMenu from './components/desktop-side-filter/store-side-menu';
 import MobileFilter from '@modules/shop/components/mobile-filter-modal/mobile-filter';
-import ProductCardGroup from '@modules/products/components/product-group-store';
+import ProductCardGroup from '@modules/products/components/product-group-shop';
 import StoreFilterDisplay from '@modules/shop/components/store-filter-display';
 
 const ShopTemplate = () => {
