+++ hamza-client/src/modules/store/components/products/store-product-display.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React, { useState, useEffect, Suspense } from 'react';
-import ProductCardGroup from '@modules/products/components/product-group-vendor';
+import ProductCardGroup from '@modules/products/components/product-group-store';
 import { Flex } from '@chakra-ui/react';
 import SearchBar from '@modules/home/components/search-and-filter-panel/components/SearchBar';
 import StoreSearch from './store-search';
